edit add decode token auth.ts

+++ intelligent-dental/src/app/api/medical_records/[id]/route.ts
@@ -8,7 +8,8 @@ export async function GET(request: Request,{ params }: { params: Promise<{ id: s
     try {
         const { id } = await params;
         const medical_id = parseInt(id);
-        const user = getCurrentUser()
+        const user = getCurrentUser(request)
+        if (!user) return res.error(401, "AUTH-001", "validation fail", "VALIDATION")
         requireRole(user.role, ["patient", "dentist"])
         const data = await medicalService.getInspectionRecordById(medical_id)
         return res.ok(data)
@@ -21,7 +22,8 @@ export async function PUT(request: Request,{ params }: { params: Promise<{ id: s
     try {
         const { id } = await params;
         const medical_id = parseInt(id);
-        const user = getCurrentUser()
+        const user = getCurrentUser(request)
+        if (!user) return res.error(401, "AUTH-001", "validation fail", "VALIDATION")
         requireRole(user.role, ["staff", "dentist"])
         const body = await request.json()
         const updatedSchedule = await medicalService.updateInspectionRecord(medical_id, body)
@@ -35,7 +37,8 @@ export async function DELETE(request: Request,{ params }: { params: Promise<{ id
     try {
         const { id } = await params;
         const medical_id = parseInt(id);
-        const user = getCurrentUser()
+        const user = getCurrentUser(request)
+        if (!user) return res.error(401, "AUTH-001", "validation fail", "VALIDATION")
         requireRole(user.role, ["staff", "dentist"])
         await medicalService.deleteInspectionRecord(medical_id)
         return res.noContent()
